fix(ci): satisfy openclaw structural gates
