+++ crates/clawdentity-core/src/providers/openclaw/cli.rs
@@ -8,11 +8,7 @@ use crate::providers::resolve_command_path;
 
 const OPENCLAW_BINARY: &str = "openclaw";
 
-fn openclaw_command(
-    command_path: &Path,
-    config_path: &Path,
-    openclaw_dir: &Path,
-) -> Command {
+fn openclaw_command(command_path: &Path, config_path: &Path, openclaw_dir: &Path) -> Command {
     let mut command = Command::new(command_path);
     command.env("OPENCLAW_CONFIG_PATH", config_path);
     command.env("OPENCLAW_STATE_DIR", openclaw_dir);
@@ -26,11 +22,7 @@ fn format_command(command_path: &Path, args: &[&str]) -> String {
     parts.join(" ")
 }
 
-fn command_error(
-    command_path: &Path,
-    args: &[&str],
-    output: &std::process::Output,
-) -> CoreError {
+fn command_error(command_path: &Path, args: &[&str], output: &std::process::Output) -> CoreError {
     let stdout = String::from_utf8_lossy(&output.stdout).trim().to_string();
     let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
     let message = if !stderr.is_empty() {
@@ -46,6 +38,7 @@ fn command_error(
     }
 }
 
+/// Resolve the `openclaw` binary used by install and setup flows.
 pub fn ensure_openclaw_cli_available(path_override: Option<&[PathBuf]>) -> Result<PathBuf> {
     resolve_command_path(OPENCLAW_BINARY, path_override).ok_or_else(|| {
         CoreError::InvalidInput(
@@ -54,6 +47,7 @@ pub fn ensure_openclaw_cli_available(path_override: Option<&[PathBuf]>) -> Resul
     })
 }
 
+/// Run `openclaw config set` with a JSON payload against the target profile.
 pub fn run_openclaw_config_set_json(
     command_path: &Path,
     config_path: &Path,
