+++ crates/clawdentity-core/src/providers/openclaw/mod.rs
@@ -187,8 +187,12 @@ impl OpenclawProvider {
         let (_, relay_snapshot_path) =
             self.resolve_setup_runtime_paths(opts, &context.config_dir, &context.openclaw_dir);
         let marker_path = write_selected_openclaw_agent(&context.config_dir, &context.agent_name)?;
-        let runtime_path =
-            self.save_setup_runtime_config(context, opts, connector_base_url, &relay_snapshot_path)?;
+        let runtime_path = self.save_setup_runtime_config(
+            context,
+            opts,
+            connector_base_url,
+            &relay_snapshot_path,
+        )?;
         let connector_assignment_path = save_connector_assignment(
             &context.config_dir,
             &context.agent_name,
@@ -226,7 +230,9 @@ impl OpenclawProvider {
     ) -> Result<PathBuf> {
         let resolved_base_url =
             resolve_openclaw_base_url(&context.config_dir, opts.platform_base_url.as_deref())?;
-        let existing_runtime = load_relay_runtime_config(&context.config_dir).ok().flatten();
+        let existing_runtime = load_relay_runtime_config(&context.config_dir)
+            .ok()
+            .flatten();
         let config_path =
             resolve_openclaw_config_path(context.state_options.home_dir.as_deref(), None)?;
         save_relay_runtime_config(
@@ -568,362 +574,7 @@ mod relay_test;
 mod setup;
 
 #[cfg(test)]
-pub(crate) mod test_support {
-    use std::fs;
-    use std::path::{Path, PathBuf};
-
-    use tempfile::TempDir;
-
-    use super::{OPENCLAW_CONFIG_FILE_NAME, resolve_openclaw_dir};
-
-    pub(crate) fn install_mock_openclaw_cli() -> TempDir {
-        let bin_dir = TempDir::new().expect("temp bin");
-        let script_path = bin_dir.path().join("openclaw");
-        let script = r#"#!/usr/bin/env node
-const fs = require("fs");
-const path = require("path");
-
-const args = process.argv.slice(2);
-const configPath = process.env.OPENCLAW_CONFIG_PATH;
-
-if (!configPath) {
-  console.error("OPENCLAW_CONFIG_PATH is required");
-  process.exit(1);
-}
-
-function readConfig() {
-  if (!fs.existsSync(configPath)) {
-    return {};
-  }
-  const raw = fs.readFileSync(configPath, "utf8").trim();
-  return raw ? JSON.parse(raw) : {};
-}
-
-function writeConfig(value) {
-  fs.mkdirSync(path.dirname(configPath), { recursive: true });
-  fs.writeFileSync(configPath, `${JSON.stringify(value, null, 2)}\n`);
-}
-
-function setPath(target, dottedPath, value) {
-  const parts = dottedPath.split(".");
-  let cursor = target;
-  for (const part of parts.slice(0, -1)) {
-    if (!cursor[part] || typeof cursor[part] !== "object" || Array.isArray(cursor[part])) {
-      cursor[part] = {};
-    }
-    cursor = cursor[part];
-  }
-  cursor[parts[parts.length - 1]] = value;
-}
-
-function getPath(target, dottedPath) {
-  return dottedPath.split(".").reduce((cursor, part) => {
-    if (cursor === undefined || cursor === null) {
-      return undefined;
-    }
-    return cursor[part];
-  }, target);
-}
-
-if (args[0] === "config" && args[1] === "set") {
-  const config = readConfig();
-  setPath(config, args[2], JSON.parse(args[3]));
-  writeConfig(config);
-  process.exit(0);
-}
-
-if (args[0] === "config" && args[1] === "get") {
-  const value = getPath(readConfig(), args[2]);
-  if (value === undefined) {
-    console.error("Config path not found");
-    process.exit(1);
-  }
-  process.stdout.write(`${JSON.stringify(value)}\n`);
-  process.exit(0);
-}
-
-console.error(`unsupported mock openclaw command: ${args.join(" ")}`);
-process.exit(1);
-"#;
-        fs::write(&script_path, script).expect("mock openclaw script");
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mut permissions = fs::metadata(&script_path).expect("metadata").permissions();
-            permissions.set_mode(0o755);
-            fs::set_permissions(&script_path, permissions).expect("chmod");
-        }
-        bin_dir
-    }
-
-    pub(crate) fn write_openclaw_profile(home_dir: &Path, config_body: &str) -> PathBuf {
-        let openclaw_dir = resolve_openclaw_dir(Some(home_dir), None).expect("openclaw dir");
-        fs::create_dir_all(&openclaw_dir).expect("openclaw dir");
-        let config_path = openclaw_dir.join(OPENCLAW_CONFIG_FILE_NAME);
-        fs::write(&config_path, config_body).expect("openclaw config");
-        config_path
-    }
-}
+pub(crate) mod test_support;
 
 #[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-    use std::fs;
-
-    use serde_json::Value;
-    use tempfile::TempDir;
-
-    use crate::{
-        config::{ConfigPathOptions, get_config_dir},
-        provider::{
-            InboundMessage, InstallOptions, PlatformProvider, ProviderDoctorOptions,
-            ProviderSetupOptions,
-        },
-    };
-
-    use super::{
-        assets::{transform_peers_path, transform_runtime_path},
-        OpenclawProvider, OPENCLAW_CONFIG_FILE_NAME, load_connector_assignments,
-        resolve_openclaw_dir,
-        test_support::{install_mock_openclaw_cli, write_openclaw_profile},
-    };
-
-    #[test]
-    fn detection_checks_home_and_path_evidence() {
-        let home = TempDir::new().expect("temp home");
-        write_openclaw_profile(home.path(), "{}\n");
-        let bin_dir = install_mock_openclaw_cli();
-
-        let provider = OpenclawProvider::with_test_context(
-            home.path().to_path_buf(),
-            vec![bin_dir.path().to_path_buf()],
-        );
-        let detection = provider.detect();
-
-        assert!(detection.detected);
-        assert!(detection.confidence > 0.9);
-        assert!(
-            detection
-                .evidence
-                .iter()
-                .any(|entry| entry.contains("openclaw binary in PATH"))
-        );
-    }
-
-    #[test]
-    fn format_inbound_uses_openclaw_webhook_shape() {
-        let provider = OpenclawProvider::default();
-        let mut metadata = HashMap::new();
-        metadata.insert("thread".to_string(), "relay".to_string());
-
-        let request = provider.format_inbound(&InboundMessage {
-            sender_did: "did:cdi:registry.clawdentity.com:agent:01HF7YAT00W6W7CM7N3W5FDXTB"
-                .to_string(),
-            recipient_did: "did:cdi:registry.clawdentity.com:agent:01HF7YAT00W6W7CM7N3W5FDXTC"
-                .to_string(),
-            content: "hello".to_string(),
-            request_id: Some("req-123".to_string()),
-            metadata,
-        });
-
-        assert_eq!(
-            request
-                .headers
-                .get("x-webhook-sender-id")
-                .map(String::as_str),
-            Some("did:cdi:registry.clawdentity.com:agent:01HF7YAT00W6W7CM7N3W5FDXTB")
-        );
-        assert_eq!(
-            request.body.get("content").and_then(|value| value.as_str()),
-            Some("hello")
-        );
-        assert_eq!(
-            request.body.get("path").and_then(|value| value.as_str()),
-            Some("/hooks/agent")
-        );
-    }
-
-    #[test]
-    fn config_path_points_to_openclaw_json() {
-        let home = TempDir::new().expect("temp home");
-        let provider = OpenclawProvider::with_test_context(home.path().to_path_buf(), Vec::new());
-
-        assert_eq!(
-            provider.config_path(),
-            Some(
-                resolve_openclaw_dir(Some(home.path()), None)
-                    .expect("openclaw dir")
-                    .join(OPENCLAW_CONFIG_FILE_NAME)
-            )
-        );
-    }
-
-    #[test]
-    fn setup_honors_explicit_connector_url_and_custom_peers_path() {
-        let home = TempDir::new().expect("temp home");
-        let bin_dir = install_mock_openclaw_cli();
-        write_openclaw_profile(
-            home.path(),
-            r#"{
-  "gateway": {
-    "auth": {
-      "mode": "password",
-      "password": "existing-password"
-    }
-  }
-}
-"#,
-        );
-        let provider = OpenclawProvider::with_test_context(
-            home.path().to_path_buf(),
-            vec![bin_dir.path().to_path_buf()],
-        );
-        let openclaw_dir = resolve_openclaw_dir(Some(home.path()), None).expect("openclaw dir");
-        let custom_peers_path = home.path().join("runtime").join("custom-peers.json");
-
-        let result = provider
-            .setup(&ProviderSetupOptions {
-                home_dir: None,
-                agent_name: Some("alpha".to_string()),
-                platform_base_url: Some("http://127.0.0.1:19001".to_string()),
-                webhook_host: None,
-                webhook_port: None,
-                webhook_token: Some("hook-token".to_string()),
-                connector_base_url: Some("https://relay.example.test:24444".to_string()),
-                connector_url: None,
-                relay_transform_peers_path: Some(custom_peers_path.display().to_string()),
-            })
-            .expect("setup");
-
-        assert!(
-            result
-                .updated_paths
-                .iter()
-                .any(|path| path == &custom_peers_path.display().to_string())
-        );
-        assert!(custom_peers_path.exists());
-        assert!(!transform_peers_path(&openclaw_dir).exists());
-
-        let runtime_path = transform_runtime_path(&openclaw_dir);
-        let runtime: Value =
-            serde_json::from_str(&fs::read_to_string(&runtime_path).expect("runtime body"))
-                .expect("runtime json");
-        assert_eq!(
-            runtime.get("connectorBaseUrl").and_then(Value::as_str),
-            Some("https://relay.example.test:24444/")
-        );
-        assert_eq!(
-            runtime
-                .get("connectorBaseUrls")
-                .and_then(Value::as_array)
-                .map(|entries| {
-                    entries
-                        .iter()
-                        .filter_map(Value::as_str)
-                        .collect::<Vec<_>>()
-                }),
-            Some(vec!["https://relay.example.test:24444/"])
-        );
-        assert_eq!(
-            runtime.get("peersConfigPath").and_then(Value::as_str),
-            Some(custom_peers_path.to_string_lossy().as_ref())
-        );
-
-        let config_dir = get_config_dir(&ConfigPathOptions {
-            home_dir: Some(home.path().to_path_buf()),
-            registry_url_hint: None,
-        })
-        .expect("config dir");
-        let assignments = load_connector_assignments(&config_dir).expect("assignments");
-        assert_eq!(
-            assignments
-                .agents
-                .get("alpha")
-                .map(|entry| entry.connector_base_url.as_str()),
-            Some("https://relay.example.test:24444/")
-        );
-
-        let config_body =
-            fs::read_to_string(openclaw_dir.join(super::OPENCLAW_CONFIG_FILE_NAME)).expect("config");
-        let config: Value = serde_json::from_str(&config_body).expect("config json");
-        assert_eq!(
-            config
-                .get("gateway")
-                .and_then(Value::as_object)
-                .and_then(|value| value.get("auth"))
-                .and_then(Value::as_object)
-                .and_then(|value| value.get("mode"))
-                .and_then(Value::as_str),
-            Some("password")
-        );
-    }
-
-    #[test]
-    fn install_requires_openclaw_cli() {
-        let home = TempDir::new().expect("temp home");
-        write_openclaw_profile(
-            home.path(),
-            r#"{
-  "gateway": {
-    "auth": {
-      "mode": "token",
-      "token": "gateway-token"
-    }
-  }
-}
-"#,
-        );
-        let provider = OpenclawProvider::with_test_context(home.path().to_path_buf(), Vec::new());
-
-        let error = provider
-            .install(&InstallOptions {
-                home_dir: Some(home.path().to_path_buf()),
-                ..InstallOptions::default()
-            })
-            .expect_err("missing openclaw CLI should fail");
-
-        assert!(error.to_string().contains("OpenClaw CLI is required"));
-    }
-
-    #[test]
-    fn doctor_does_not_require_openclaw_cli() {
-        let home = TempDir::new().expect("temp home");
-        let provider = OpenclawProvider::with_test_context(home.path().to_path_buf(), Vec::new());
-
-        let result = provider
-            .doctor(&ProviderDoctorOptions {
-                home_dir: Some(home.path().to_path_buf()),
-                include_connector_runtime_check: false,
-                ..ProviderDoctorOptions::default()
-            })
-            .expect("doctor should inspect local state without the openclaw CLI");
-
-        assert_eq!(
-            result
-                .checks
-                .iter()
-                .find(|check| check.id == "state.openclawConfig")
-                .map(|check| check.message.as_str()),
-            Some("OpenClaw config is missing")
-        );
-    }
-
-    #[test]
-    fn setup_requires_openclaw_onboarding_first() {
-        let home = TempDir::new().expect("temp home");
-        let bin_dir = install_mock_openclaw_cli();
-        let provider = OpenclawProvider::with_test_context(
-            home.path().to_path_buf(),
-            vec![bin_dir.path().to_path_buf()],
-        );
-
-        let error = provider
-            .setup(&ProviderSetupOptions {
-                agent_name: Some("alpha".to_string()),
-                ..ProviderSetupOptions::default()
-            })
-            .expect_err("missing openclaw config should fail");
-
-        assert!(error.to_string().contains("openclaw onboard"));
-    }
-}
+mod mod_tests;
